requirements.txt
asyncinotify==3.1.2
b2sdk==1.19.0
docopt==0.6.2
fastapi[all]==0.89.0
mutagen==1.46.0
# mysql-connector-python==8.0.26
# mysqlclient==2.0.3
SQLAlchemy==1.4.23
libtorrent==2.0.7
aio-pika==8.3.0
PyMySQL==1.0.2
cryptography==39.0.0
pydub==0.25.1
Pillow==9.4.0
ImageHash==4.3.1
python-jose[cryptography]
passlib[bcrypt]